backend/requirements.txt
django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pdfplumber==0.11.4
pydantic==2.8.2
temporalio==1.7.1
httpx==0.27.2
gunicorn==22.0.0
whitenoise==6.7.0
